Check for dax file before deleting in steve/capture

+++ steve/capture.py
@@ -176,7 +176,9 @@ class Capture(QtCore.QObject):
     #
     @hdebug.debug
     def captureStart(self, stagex, stagey):
-        os.remove(self.directory + self.filename + ".dax")
+        ## HBCheck:  Should we check for the file before deleting it?
+        if os.path.exists(self.directory + self.filename + ".dax"):
+            os.remove(self.directory + self.filename + ".dax")
         
         if not self.tcp_client.isConnected():
             hdebug.logText("captureStart: not connected to HAL.")
